refactor: reorganize components and improve naming
- Move context providers to dedicated contexts directory
- Rename ETF components to timelock for better clarity
- Update import paths across the application
