+++ src/app/compose/page.tsx
@@ -25,8 +25,8 @@ import { Input, InputGroup } from '@/components/input'
 import { Link } from '@/components/link'
 import { Select } from '@/components/select'
 import { EllipsisVerticalIcon, MagnifyingGlassIcon } from '@heroicons/react/16/solid'
-import { NUMBER_BLOCKS_EXECUTED, useConnectedWallet } from '@/components/etf/connectedWalletContext'
-import { ConnectWallet } from '@/components/etf/connectWallet'
+import { NUMBER_BLOCKS_EXECUTED, useConnectedWallet } from '@/components/contexts/connectedWalletContext'
+import { ConnectWallet } from '@/components/timelock/connectWallet'
 import { useState } from 'react'
 import { ExclamationTriangleIcon, XCircleIcon } from '@heroicons/react/20/solid'
 import { explorerClient } from '@/lib/explorer-client';
