+++ src/app/compose/schedule/page.tsx
@@ -24,9 +24,9 @@ import { Link } from '@/components/link'
 import { Input } from '@/components/input'
 import { Heading, Subheading } from '@/components/heading'
 import { ChevronDoubleLeftIcon } from '@heroicons/react/16/solid'
-import { DynamicExtrinsicForm } from '@/components/etf/dynamicExtrinsicForm'
-import { useConnectedWallet } from '@/components/etf/connectedWalletContext'
-import { ConnectWallet } from '@/components/etf/connectWallet'
+import { DynamicExtrinsicForm } from '@/components/timelock/dynamicExtrinsicForm'
+import { useConnectedWallet } from '@/components/contexts/connectedWalletContext'
+import { ConnectWallet } from '@/components/timelock/connectWallet'
 import { useState } from 'react'
 import { DelayedTransactionDetails } from '@/domain/DelayedTransactionDetails'
 import { XCircleIcon } from '@heroicons/react/20/solid'
