+++ src/components/timelock/connectWallet.tsx
@@ -36,7 +36,7 @@ import { Button } from "@/components/button";
 import { Dialog, DialogTitle } from "@/components/dialog";
 import { Text } from "@/components/text";
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/table";
-import { useConnectedWallet } from "@/components/etf/connectedWalletContext";
+import { useConnectedWallet } from "@/components/contexts/connectedWalletContext";
 import { ArrowRightStartOnRectangleIcon } from "@heroicons/react/20/solid";
 import { explorerClient } from '@/lib/explorer-client';
